Homework 4
Create a class according to the specification. Define a constructor for all fields. Create a separate processor class that will perform certain actions with objects and contain a method for printing an object. In the main method, create an array of objects and use the processor class to perform the necessary operations.

Car: id, Brand, Model, Year of manufacture, Color, Price, Registration number.

Create an array of objects. Output:

a) a list of cars of a given brand;

b) a list of cars of a given model that have been in operation for more than n years;

c) a list of cars of a given year of manufacture, the price of which is higher than the specified